mv getDataToCoordFunc in axes.js
- use it filter and sort modules.
